rename DocumentDatabase to RootDatabase
