+++ crates/apollo-compiler/src/queries/mod.rs
@@ -1,10 +1,13 @@
 pub mod database;
 pub mod values;
 
-mod definitions_db;
+mod def_db;
+mod document_db;
 mod inputs_db;
 mod parser_db;
 
-pub use definitions_db::Definitions;
+pub use database::RootDatabase;
+pub use def_db::Definitions;
+pub use document_db::Document;
 pub use inputs_db::Inputs;
 pub use parser_db::DocumentParser;
